chore(signature): remove envelope module
