+++ backend/signature/urls.py
@@ -1,8 +1,8 @@
 # signature/urls.py
 from rest_framework.routers import DefaultRouter
 from django.urls import path, include
-from .views.envelope import (EnvelopeViewSet,PrintQRCodeViewSet,guest_envelope_view,serve_decrypted_pdf)
-from .views.auth import (register,activate_account,user_profile,password_reset_request,change_password,)
+from .views.recipient_view import EnvelopeViewSet, PrintQRCodeViewSet
+from .views.guest_view import guest_envelope_view, serve_decrypted_pdf
 from .views.notification import NotificationPreferenceViewSet
 from .views.batch import SelfSignView, BatchSignCreateView, BatchSignJobViewSet
 from .views.saved_signature import SavedSignatureViewSet
